@@ -1,8 +1,7 @@
-namespace WebApp_OpenIDConnect_DotNet.Infrastructure
+namespace WebApp_OpenIDConnect_DotNet.Infrastructure;
+
+public static class Constants
 {
-    public static class Constants
-    {
-        public const string ScopeUserImpersonation = "user_impersonation";
-        public const string BearerAuthorizationScheme = "Bearer";
-    }
+    public const string ScopeUserImpersonation = "user_impersonation";
+    public const string BearerAuthorizationScheme = "Bearer";
 }
