@@ -1,9 +1,8 @@
-namespace WebApp_OpenIDConnect_DotNet.Models
-{
-    public class ErrorViewModel
-    {
-        public string RequestId { get; set; }
+namespace WebApp_OpenIDConnect_DotNet.Models;
 
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-    }
+public class ErrorViewModel
+{
+    public string RequestId { get; set; }
+
+    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
 }
