@@ -1,23 +1,18 @@
 ﻿using Microsoft.AspNetCore.Http;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text.Json;
-using System.Threading.Tasks;
 
-namespace TodoListService.Common
+namespace TodoListService.Common;
+
+public static class SessionExtensions
 {
-    public static class SessionExtensions
+    public static void Set<T>(this ISession session, string key, T value)
     {
-        public static void Set<T>(this ISession session, string key, T value)
-        {
-            session.SetString(key, JsonSerializer.Serialize(value));
-        }
+        session.SetString(key, JsonSerializer.Serialize(value));
+    }
 
-        public static T Get<T>(this ISession session, string key)
-        {
-            var value = session.GetString(key);
-            return value == null ? default : JsonSerializer.Deserialize<T>(value);
-        }
+    public static T Get<T>(this ISession session, string key)
+    {
+        var value = session.GetString(key);
+        return value == null ? default : JsonSerializer.Deserialize<T>(value);
     }
 }
