@@ -44,7 +44,7 @@
         @RenderBody()
         <hr />
         <footer>
-            <p>&copy; 2018 -  WebApp_OpenIDConnect_DotNet</p>
+            <p>&copy; 2022 -  WebApp_OpenIDConnect_DotNet</p>
         </footer>
     </div>
 
