@@ -45,7 +45,7 @@
         @RenderBody()
         <hr />
         <footer>
-            <p>&copy; 2019 - WebAPI_DotNet</p>
+            <p>&copy; 2022 - WebAPI_DotNet</p>
         </footer>
     </div>
 
